Revert back to before xlabel 'fix'
The previous 'fix' to prevent the xlabel from disappearing was
to remove the code that set the label positions back to their
original position after substituting in replacements.  I had
forgotten why they were in there... for things like ylabels
which can move when a different width yticklabel is substituted
in.